refactor: LeaderDescriptionField 분리
